GPencil: Split transform code
This commit splits the grease pencil transform code, so that editcurves and strokes get handled separately.
This fixes a bug where in normal edit mode, proportional editing would be very slow.
The center point of the transform is also more predictable and behaves similar to bezier triples in curve objects.